Allow for querying for question types available for a topic
In order to assert that we can only get one question types that have valid calculators for a topic we need to only retrieve topics with valid calculators for the type. In addition to this we have certain question types that can apply to any topic. For this we added a generic column to the question type table. Rows with generic = TRUE are types that work for any question type.