Create framework for scoring choices
In order to score choices that have subjects we will loop over all games happening on the current day. From here we update the game information, and then update the scores for all choices relating to the game and all subjects it references. Subjects now can return an array of all related subjects which are then used to load the choices referencing that subject and using the question from the choice it scores them according to the scorer object that the question returns.